add gjk filtering to height field midphase
