+++ mujoco_warp/_src/collision_hfield.py
@@ -17,14 +17,9 @@ from typing import Tuple
 
 import warp as wp
 
-from .types import MJ_MAXVAL
-from .types import Data
-from .types import GeomType
-from .types import Model
-
 
 @wp.func
-def _hfield_subgrid(
+def hfield_subgrid(
   # In:
   nrow: int,
   ncol: int,
@@ -184,218 +179,3 @@ def hfield_prism_vertex(prism: wp.mat33, vert_index: int) -> wp.vec3:
       return wp.vec3(prism[0][0], prism[1][1], prism[2][2])
     else:  # odd triangle
       return wp.vec3(prism[1][0], prism[0][1], prism[2][2])
-
-
-@wp.kernel
-def _hfield_midphase(
-  # Model:
-  geom_type: wp.array(dtype=int),
-  geom_dataid: wp.array(dtype=int),
-  geom_aabb: wp.array2d(dtype=wp.vec3),
-  geom_rbound: wp.array2d(dtype=float),
-  geom_margin: wp.array2d(dtype=float),
-  hfield_nrow: wp.array(dtype=int),
-  hfield_ncol: wp.array(dtype=int),
-  hfield_size: wp.array(dtype=wp.vec4),
-  # Data in:
-  nconmax_in: int,
-  geom_xpos_in: wp.array2d(dtype=wp.vec3),
-  geom_xmat_in: wp.array2d(dtype=wp.mat33),
-  collision_pair_in: wp.array(dtype=wp.vec2i),
-  collision_hftri_index_in: wp.array(dtype=int),
-  collision_pairid_in: wp.array(dtype=int),
-  collision_worldid_in: wp.array(dtype=int),
-  # Data out:
-  collision_pair_out: wp.array(dtype=wp.vec2i),
-  collision_hftri_index_out: wp.array(dtype=int),
-  collision_pairid_out: wp.array(dtype=int),
-  collision_worldid_out: wp.array(dtype=int),
-  ncollision_out: wp.array(dtype=int),
-):
-  """Midphase collision detection for heightfield triangles with other geoms.
-
-  This kernel processes collision pairs where one geom is a heightfield (identified by
-  collision_hftri_index_in[pairid] == -1) and expands them into multiple collision pairs,
-  one for each potentially colliding triangle.
-
-  Args:
-    geom_type: geom type
-    geom_dataid: geom data id
-    geom_rbound: geom bounding sphere radius
-    geom_margin: geom margin
-    hfield_nrow: height field number of rows
-    hfield_ncol: height field number of columns
-    hfield_size: height field size
-    nconmax_in: maximum number of contacts
-    geom_xpos_in: geom position
-    geom_xmat_in: geom orientation
-    collision_pair_in: collision pair
-    collision_hftri_index_in: triangle indices, -1 for height field pair
-    collision_pairid_in: collision pair id from broadphase
-    collision_worldid_in: collision world id from broadphase
-    collision_pair_out: collision pair from midphase
-    collision_hftri_index_out: triangle indices from midphase
-    collision_pairid_out: collision pair id from midphase
-    collision_worldid_out: collision world id from midphase
-    ncollision_out: number of collisions from broadphase and midphase
-  """
-  pairid = wp.tid()
-
-  # only process pairs that are marked for height field collision (-1)
-  if collision_hftri_index_in[pairid] != -1:
-    return
-
-  # collision pair info
-  worldid = collision_worldid_in[pairid]
-  pair_id = collision_pairid_in[pairid]
-
-  pair = collision_pair_in[pairid]
-  g1 = pair[0]
-  g2 = pair[1]
-
-  hfieldid = g1
-  geomid = g2
-
-  # SHOULD NOT OCCUR: if the first geom is not a heightfield, swap
-  if geom_type[g1] != int(GeomType.HFIELD.value):
-    hfieldid = g2
-    geomid = g1
-
-  # height field info
-  hfdataid = geom_dataid[hfieldid]
-  size1 = hfield_size[hfdataid]
-  pos1 = geom_xpos_in[worldid, hfieldid]
-  mat1 = geom_xmat_in[worldid, hfieldid]
-  mat1T = wp.transpose(mat1)
-
-  # geom info
-  pos2 = geom_xpos_in[worldid, geomid]
-  pos = mat1T @ (pos2 - pos1)
-  r2 = geom_rbound[worldid, geomid]
-
-  # TODO(team): margin?
-  margin = wp.max(geom_margin[worldid, hfieldid], geom_margin[worldid, geomid])
-
-  # box-sphere test: horizontal plane
-  for i in range(2):
-    if (size1[i] < pos[i] - r2 - margin) or (-size1[i] > pos[i] + r2 + margin):
-      return
-
-  # box-sphere test: vertical direction
-  if size1[2] < pos[2] - r2 - margin:  # up
-    return
-
-  if -size1[3] > pos[2] + r2 + margin:  # down
-    return
-
-  mat2 = geom_xmat_in[worldid, geomid]
-  mat = mat1T @ mat2
-
-  # aabb for geom in height field frame
-  xmax = -MJ_MAXVAL
-  ymax = -MJ_MAXVAL
-  zmax = -MJ_MAXVAL
-  xmin = MJ_MAXVAL
-  ymin = MJ_MAXVAL
-  zmin = MJ_MAXVAL
-
-  center2 = geom_aabb[geomid, 0]
-  size2 = geom_aabb[geomid, 1]
-
-  pos += mat1T @ center2
-
-  sign = wp.vec2(-1.0, 1.0)
-
-  for i in range(2):
-    for j in range(2):
-      for k in range(2):
-        corner_local = wp.vec3(sign[i] * size2[0], sign[j] * size2[1], sign[k] * size2[2])
-        corner_hf = mat @ corner_local
-
-        if corner_hf[0] > xmax:
-          xmax = corner_hf[0]
-        if corner_hf[1] > ymax:
-          ymax = corner_hf[1]
-        if corner_hf[2] > zmax:
-          zmax = corner_hf[2]
-        if corner_hf[0] < xmin:
-          xmin = corner_hf[0]
-        if corner_hf[1] < ymin:
-          ymin = corner_hf[1]
-        if corner_hf[2] < zmin:
-          zmin = corner_hf[2]
-
-  xmax += pos[0]
-  xmin += pos[0]
-  ymax += pos[1]
-  ymin += pos[1]
-  zmax += pos[2]
-  zmin += pos[2]
-
-  # box-box test
-  if (
-    (xmin - margin > size1[0])
-    or (xmax + margin < -size1[0])
-    or (ymin - margin > size1[1])
-    or (ymax + margin < -size1[1])
-    or (zmin - margin > size1[2])
-    or (zmax + margin < -size1[3])
-  ):
-    return
-
-  # height field subgrid
-  nrow = hfield_nrow[hfieldid]
-  ncol = hfield_ncol[hfieldid]
-  size = hfield_size[hfieldid]
-  cmin, rmin, cmax, rmax = _hfield_subgrid(nrow, ncol, size, xmax, xmin, ymax, ymin)
-
-  # loop over subgrid triangles
-  for r in range(rmin, rmax):
-    for c in range(cmin, cmax):
-      # add both triangles from this cell
-      for i in range(2):
-        if r == rmin and c == cmin and i == 0:
-          # reuse the initial pair for the 1st triangle
-          new_pairid = pairid
-        else:
-          # create a new pair
-          new_pairid = wp.atomic_add(ncollision_out, 0, 1)
-
-        if new_pairid >= nconmax_in:
-          return
-
-        collision_pair_out[new_pairid] = pair
-        collision_hftri_index_out[new_pairid] = 2 * (r * (ncol - 1) + c) + i
-        collision_pairid_out[new_pairid] = pair_id
-        collision_worldid_out[new_pairid] = worldid
-
-
-def hfield_midphase(m: Model, d: Data):
-  """Midphase collision detection for height field triangles with other geoms.
-
-  Processes collision pairs from the broadphase where one geom is a height field and expands
-  them into multiple collision pairs, one for each potentially colliding triangle. The
-  function directly writes to the same collision buffers used by _add_geom_pair.
-  """
-  wp.launch(
-    kernel=_hfield_midphase,
-    dim=d.nconmax,
-    inputs=[
-      m.geom_type,
-      m.geom_dataid,
-      m.geom_aabb,
-      m.geom_rbound,
-      m.geom_margin,
-      m.hfield_nrow,
-      m.hfield_ncol,
-      m.hfield_size,
-      d.nconmax,
-      d.geom_xpos,
-      d.geom_xmat,
-      d.collision_pair,
-      d.collision_hftri_index,
-      d.collision_pairid,
-      d.collision_worldid,
-    ],
-    outputs=[d.collision_pair, d.collision_hftri_index, d.collision_pairid, d.collision_worldid, d.ncollision],
-  )
